Initial Beta v0.4.9.5
Using json to stringify an ArrayList contaning the times of each timer, then saving as string in sharedpreferences.
For some reason the JSON string is not being saved properly, and therefore is not recovered during onCreate.
It turns out that doing separate .apply() for the string and the leaving time was for some reason impeding the recuperation of the string.
Testing saving arraylist of all timers as json to keep all values like initial time, etc.