Initial Beta v0.6
Service not working properly. Going to change the app so that it uses native saving state for background counting, and stores the original times for timers when exiting the app. Appears to be a good compromise.
First version, service eliminated, saving back to sharedpreferences, will try to override back button to prevent exiting app and losing timers/progress.