Initial Beta v0.7.2
Lots of changes! Turns out the button had to be "long pressed" to exit the app, otherwise it goes to "theater or ambient mode" (at least on emulator). Also, disabled dismissing the app with back swipe. This caused many activities to overlap each time I opened the app, so the manifest was modified to allow only 1 activity at a time. The back swipe is disabled through styles and changing the manifest. Countdowntimer appears to be working just fine, even in background. Will test more...