Initial Beta v0.8.4
General cleanup of commented // unnecesary code and variables

Ambient mode implemented.

Using wakelocks to keep CPU working  while on ambient mode and a timer is running, so that it refreshes every second, as it should.

Saving implemented. Now exiting the app with a back swipe saves the initial state of timers, later restored during oncreate.

Disabled antialiasing on textview that displays the time when on ambient mode...

Future work: tutorials for the app on main screen on first start, some way to show it on demand?
                       animations for the transitions.
                       add time of day to the top when entering ambient mode?
                       add stopwatch one day...